requirements.txt
astroid==2.15.5
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pyaml==23.5.9
pylint==2.17.4
PyYAML==6.0
requests==2.31.0
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
urllib3==1.26.16
wrapt==1.15.0
